feat: moving sequelize config and  directories to database dir
